move patch related API to api/patcher package
