+++ src/main/java/org/betonquest/betonquest/api/config/patcher/PatchException.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.api.config;
+package org.betonquest.betonquest.api.config.patcher;
 
 import java.io.Serial;
