+++ src/main/java/org/betonquest/betonquest/modules/config/ConfigurationFileImpl.java
@@ -4,7 +4,7 @@ import lombok.CustomLog;
 import org.betonquest.betonquest.api.bukkit.config.custom.ConfigurationSectionDecorator;
 import org.betonquest.betonquest.api.config.ConfigAccessor;
 import org.betonquest.betonquest.api.config.ConfigurationFile;
-import org.betonquest.betonquest.api.config.Patcher;
+import org.betonquest.betonquest.api.config.patcher.Patcher;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.InvalidConfigurationException;
