+++ src/main/java/org/betonquest/betonquest/modules/config/transformers/ListEntryAddTransformation.java
@@ -1,8 +1,8 @@
 package org.betonquest.betonquest.modules.config.transformers;
 
 import lombok.CustomLog;
-import org.betonquest.betonquest.api.config.PatchException;
-import org.betonquest.betonquest.api.config.PatchTransformation;
+import org.betonquest.betonquest.api.config.patcher.PatchException;
+import org.betonquest.betonquest.api.config.patcher.PatchTransformation;
 import org.bukkit.configuration.ConfigurationSection;
 
 import java.util.List;
